+++ packages/AreaChart/src/AreaChart.stories.tsx
@@ -284,3 +284,18 @@ export const EmptyData = () => {
     </Section>
   );
 };
+
+export const Loading = () => {
+  return (
+    <AreaChart
+      loading
+      scoreFormatter={nFormatter}
+      dateRange={{
+        start: '2022-10-01',
+        end: '2022-10-07',
+      }}
+      range={{ min: 0, max: 5 }}
+      dataSet={[]}
+    />
+  );
+};