+++ website/next.config.js
@@ -28,6 +28,7 @@ const withTM = require('next-transpile-modules')([
   '@igloo-ui/toggle',
   '@igloo-ui/tooltip',
   '@igloo-ui/tag',
+  '@igloo-ui/tag-picker',
   '@igloo-ui/tabs',
   '@igloo-ui/popover',
   '@igloo-ui/modal',